requirements.txt
transformers==4.45.0
datasets
deepspeed
sentencepiece
tensorboardX
accelerate

einops
importlib-resources
jsonlines
numexpr
omegaconf
pycountry
pytablewriter
rouge-score
sacrebleu
scikit-learn
sqlitedict
tqdm-multiprocess
zstandard

bitsandbytes
peft
fraction